llama-server-multimodal/gradio_requirements.txt
gradio>=4.0.0
openai>=1.0.0
python-dotenv>=1.0.0
